+++ package.json
@@ -13,6 +13,7 @@
     "@fullcalendar/timegrid": "^5.10.1",
     "@microsoft/applicationinsights-react-js": "3.2.1",
     "@microsoft/applicationinsights-web": "2.7.1",
+    "patch-package": "^6.4.7",
     "react": "17.0.2",
     "react-big-calendar": "0.38.0",
     "react-bootstrap": "^1.6.4",
@@ -26,7 +27,8 @@
     "start": "cross-env BROWSER=none react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "postinstall": "patch-package"
   },
   "eslintConfig": {
     "extends": [
