+++ src/index.css
@@ -130,10 +130,14 @@
         display: block;
         margin: 0.2vh;
     }
+
+    .fc .fc-header-toolbar .fc-toolbar-title {
+        margin: 1vh;
+    }
 }
 
-@media (max-height: 750px) {
-    /* .navbar.fixed-bottom {
-        visibility: hidden;
-    } */
+.fc .fc-timegrid-slots .fc-timegrid-slot {
+    height: 5vh;
+    border-bottom: 0 !important;
+    vertical-align: top;
 }
